requirements.txt
paho-mqtt
pyqt5